add not found error when user is deleted

+++ backend/app/Controllers/Http/UserController.js
@@ -4,7 +4,7 @@
 const User = use('App/Models/User');
 
 class UserController {
-  async store({ request }) {
+  async store({ request, response }) {
     const { username, email, password } = request.only([
       'username',
       'email',
@@ -13,13 +13,17 @@ class UserController {
 
     const user = await User.create({ username, email, password });
 
-    return user;
+    return response.created(user);
   }
 
   async destroy({ params, response }) {
     const { id } = params;
 
     const user = await User.find(id);
+
+    if (user === null) {
+      return response.status(404).send({ error: 'user not found' });
+    }
 
     await user.delete();
   }
